requirements.txt
flask==3.0.0
flask-cors==4.0.0
pandas==2.1.3
openpyxl==3.1.2
PyPDF2==3.0.1
python-docx==1.1.0
werkzeug==3.0.1
python-dotenv==1.0.0
openai==1.12.0

